pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "desktop-monitor"
version = "0.1.0"
description = "AI-powered desktop activity monitor with Ollama LLM integration"
requires-python = ">=3.11"
dependencies = [
    "mss>=9.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "pygetwindow>=0.0.9",
    "psutil>=5.9",
    "pyperclip>=1.8",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "httpx>=0.27",
    "apscheduler>=3.10",
    "aiosqlite>=0.20",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
desktop-monitor = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
